+++ lib/segment/src/index/field_index/numeric_index/builders.rs
@@ -10,8 +10,8 @@ use serde_json::Value;
 use super::Encodable;
 use super::mmap_numeric_index::MmapNumericIndex;
 use super::mutable_numeric_index::InMemoryNumericIndex;
+use super::numeric_index::{NumericIndex, NumericIndexIntoInnerValue};
 use super::storage::NumericIndexInner;
-use super::wrapper::{NumericIndex, NumericIndexIntoInnerValue};
 use crate::common::operation_error::{OperationError, OperationResult};
 use crate::index::field_index::numeric_point::Numericable;
 use crate::index::field_index::stored_point_to_values::StoredValue;
