+++ lib/segment/src/index/field_index/numeric_index/value_indexer.rs
@@ -5,8 +5,8 @@ use common::types::PointOffsetType;
 use serde_json::{Number, Value};
 use uuid::Uuid;
 
+use super::numeric_index::{NumericIndex, NumericIndexIntoInnerValue};
 use super::storage::NumericIndexInner;
-use super::wrapper::{NumericIndex, NumericIndexIntoInnerValue};
 use crate::common::operation_error::{OperationError, OperationResult};
 use crate::common::utils::MultiValue;
 use crate::index::field_index::ValueIndexer;
